requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==5.0.6
djangorestframework==3.15.1
drf-yasg==1.21.7
Faker==25.2.0
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
openapi-codec==1.3.2
packaging==24.0
pluggy==1.5.0
pytest==8.2.1
pytest-django==4.8.0
pytest-progress==1.2.5
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
setuptools==70.0.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
